doctor/doctorapp/requirements.txt
fastapi
uvicorn
psycopg2-binary
jinja2
python-multipart
python-dotenv
httpx
